refactor(tui): separate focus map inspection
